# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Addis Mesob Property Management System (AM-PMS) API
  version: 1.0.0
  description: Backend API for AM-PMS. Phase 1 implements Auth, RBAC, User Management, Role Management, and Append-Only Audit Logging.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
...
paths:
...
  /branches:
    get:
      tags:
        - Master Data - Branches
      summary: List branches
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of branches
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Branch'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Master Data - Branches
      summary: Create Branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBranchInput'
      responses:
        '201':
          description: Branch created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Branch'
        '400':
          description: Validation failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /branches/{id}:
    get:
      tags:
        - Master Data - Branches
      summary: Get Branch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Branch details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Branch'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    patch:
      tags:
        - Master Data - Branches
      summary: Update Branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBranchInput'
      responses:
        '200':
          description: Branch updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Branch'
        '400':
          description: Validation failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    delete:
      tags:
        - Master Data - Branches
      summary: Hard delete Branch (Blocked per FR-MD-05)
      description: Hard deletes are prohibited on master data entities per FR-MD-05. Always returns 405 Method Not Allowed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /branches/{id}/deactivate:
    post:
      tags:
        - Master Data - Branches
      summary: Deactivate Branch (Soft delete per FR-MD-05)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Branch deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Branch'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
  /branches/{id}/history:
    get:
      tags:
        - Master Data - Branches
      summary: Get Branch change history (FR-MD-06)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Version and change history list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MasterDataHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /buildings:
    get:
      tags:
        - Master Data - Buildings
      summary: List buildings
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Building'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Master Data - Buildings
      summary: Create Building
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingInput'
      responses:
        '201':
          description: Building created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Building'
        '400':
          description: Validation failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /buildings/{id}:
    get:
      tags:
        - Master Data - Buildings
      summary: Get Building by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Building details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Building'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    patch:
      tags:
        - Master Data - Buildings
      summary: Update Building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBuildingInput'
      responses:
        '200':
          description: Building updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Building'
        '400':
          description: Validation failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    delete:
      tags:
        - Master Data - Buildings
      summary: Hard delete Building (Blocked per FR-MD-05)
      description: Hard deletes are prohibited on master data entities per FR-MD-05. Always returns 405 Method Not Allowed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /buildings/{id}/deactivate:
    post:
      tags:
        - Master Data - Buildings
      summary: Deactivate Building (Soft delete per FR-MD-05)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Building deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Building'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
  /buildings/{id}/history:
    get:
      tags:
        - Master Data - Buildings
      summary: Get Building change history (FR-MD-06)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Version and change history list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MasterDataHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /floors:
    get:
      tags:
        - Master Data - Floors
      summary: List floors
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of floors
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Floor'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Master Data - Floors
      summary: Create Floor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFloorInput'
      responses:
        '201':
          description: Floor created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Floor'
        '400':
          description: Validation failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /floors/{id}:
    get:
      tags:
        - Master Data - Floors
      summary: Get Floor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Floor details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Floor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    patch:
      tags:
        - Master Data - Floors
      summary: Update Floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFloorInput'
      responses:
        '200':
          description: Floor updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Floor'
        '400':
          description: Validation failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    delete:
      tags:
        - Master Data - Floors
      summary: Hard delete Floor (Blocked per FR-MD-05)
      description: Hard deletes are prohibited on master data entities per FR-MD-05. Always returns 405 Method Not Allowed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /floors/{id}/deactivate:
    post:
      tags:
        - Master Data - Floors
      summary: Deactivate Floor (Soft delete per FR-MD-05)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Floor deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Floor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    Branch:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        nameAm:
          type: string
        code:
          type: string
        address:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
...
    UpdateBranchInput:
      type: object
      properties:
        name:
          type: string
        nameAm:
          type: string
        code:
          type: string
        address:
          type: string
...
  responses:
    Unauthorized:
      description: Unauthorized or token invalid/expired
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: UNAUTHORIZED
                  message:
                    type: string
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
...
security:
  - BearerAuth: []